requirements.txt
fastapi
pydantic
openai
python-dotenv
sentence-transformers
psycopg2
requests
beautifulsoup4
concurrent
langchain
logging
